+++ model/source.hpp
@@ -58,7 +58,7 @@ public:
      *
      * @return l'angle du rayon émis.
      */
-    int getAngle() const;
+    double getAngle() const;
 
     /**
      * Retourne la longueur du côté du carré.
